refactor: extract pagination logic to base class
