+++ src/modules/star-system/dto/inputs/pagination-star-system.dto.ts
@@ -1,22 +1,3 @@
-import { IsIn, IsInt, IsOptional, Min } from 'class-validator';
-import { Type } from 'class-transformer';
+import { Pagination } from '@shared/interceptors';
 
-export type OrderBy = 'asc' | 'desc';
-
-export class StarSystemPagination {
-  @IsOptional()
-  @Type(() => Number)
-  @IsInt()
-  @Min(1)
-  page: number = 1;
-
-  @IsOptional()
-  @Type(() => Number)
-  @IsInt()
-  @Min(1)
-  limit: number = 10;
-
-  @IsOptional()
-  @IsIn(['asc', 'desc'])
-  orderBy: OrderBy = 'asc';
-}
+export class StarSystemPagination extends Pagination {}
